revisado os endpoints existentes de reserva, feito a verificação de exclusão de salas

+++ back-end/src/main/java/br/edu/ifpr/irati/ads/agenda_inteligente/service/ClassroomService.java
@@ -66,6 +66,10 @@ public class ClassroomService {
     public boolean deleteClassroom(String id) {
         Optional<Classroom> optionalClassroom = classroomRepository.findById(id);
 
+        if (classroomRepository.existsByIdAndReservationsIsNotEmpty(id)) {
+            return false;
+        };
+
         if (optionalClassroom.isPresent()) {
             classroomRepository.delete(optionalClassroom.get());
